% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_toy_pldata}
\alias{generate_toy_pldata}
\title{A function to generate toy partial linear model data}
\usage{
generate_toy_pldata()
}
\description{
A function to generate toy partial linear model data
}
