% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_toy_pldata}
\alias{generate_toy_pldata}
\title{A function to generate toy partial linear model data}
\usage{
generate_toy_pldata()
}
\value{
a list that contains:
\itemize{
\item ysample - the response variable
\item x_cov - the linear predictors
\item times - the non-linear term
}
}
\description{
A function to generate toy partial linear model data
}
